improve filter for zoneId
if the filter param to findServices has a zoneId property that is a string
array then these strings will be matched against the personal zone id,
which is the openid user account.

previously that zoneId could be a partial match. with this commit this
matches only if a string in the zoneId filter propertry is equal to the
zone id.

Example:
// will match services from zone foo@bar.baz
var filter = { zoneId: ['foo@bar.baz'] };

Jira-issue: WP-783